New horizontal pose and isolation of arm's function in arm_scripts folder which are called through a module using rospack directory
